lexer: Add private data into Token struct
The definition of make_* functions are similar, so consolidate
these functions to a single `make_token()` and use define to keep
origin make_* function exist for readability.

To achieve this goal, we have to merge the union inside `struct Token`.
Each of the elements in the union is relative to one or more Token enum types,
but actully they can be merged to a single private data.

And the key is choosing the variable type of private data, `uintptr_t` is
sufficient to both 32 and 64-bit system for most case, however, if a
32-bit system adds a new TokenType and stores 64-bit wide private data,
the uintptr_t would be insufficient in this case.