Make the user interface to be more like gcc.
After this change, the user does not need to redirect the stdin and
stdout. They can use MazuCC just like gcc. To be more accurately,
before this change, we have to redirect the stdin and stdout :

              ./mzcc < input.c > output.S

which is not intuitive, and not what we are used to.
After this change, we can use MazuCC just like gcc :

              ./mzcc -o output.S input.c